requirement.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.1.3
django-ckeditor==6.0.0
django-comments-dab==2.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-swagger==2.0.7
djangorestframework==3.12.2
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
kavenegar==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.1.2
psycopg2==2.8.5
pylint==2.6.0
python3-openid==3.2.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.9.0
sqlparse==0.3.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==16.7.9
Werkzeug==1.0.1
wrapt==1.12.1
